Improvements and Bug Fixes
Made it more efficient (although it doesn't really matter for such a
simple program) and made it follow the correct Mastermind procedures:
"If there are duplicate colours in the guess, they cannot all be awarded
a key peg unless they correspond to the same number of duplicate colours
in the hidden code. For example, if the hidden code is
white-white-black-black and the player guesses white-white-white-black,
the codemaker will award two colored key pegs for the two correct
whites, nothing for the third white as there is not a third white in the
code, and a colored key peg for the black. No indication is given of the
fact that the code also includes a second black"